backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
Pillow==10.4.0
openpyxl==3.1.5
celery==5.4.0
redis==5.1.1
django-celery-beat==2.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
python-decouple==3.8
gunicorn==23.0.0
django-storages==1.14.4
boto3==1.35.36
